+++ polars/polars-core/src/chunked_array/kernels/mod.rs
@@ -1,8 +1,1 @@
 pub(crate) mod take;
-#[cfg(any(
-    feature = "temporal",
-    feature = "dtype-datetime",
-    feature = "dtype-date"
-))]
-#[cfg_attr(docsrs, doc(cfg(feature = "temporal")))]
-pub mod temporal;
