+++ polars/polars-core/src/chunked_array/temporal/datetime.rs
@@ -5,7 +5,6 @@ use crate::prelude::*;
 use arrow::temporal_conversions::{
     timestamp_ms_to_datetime, timestamp_ns_to_datetime, timestamp_us_to_datetime,
 };
-use std::fmt::Write;
 
 impl DatetimeChunked {
     pub fn as_datetime_iter(
