+++ polars/polars-core/src/chunked_array/temporal/time.rs
@@ -1,11 +1,7 @@
 use super::*;
-use crate::chunked_array::kernels::temporal::{
-    time_to_hour, time_to_minute, time_to_nanosecond, time_to_second,
-};
 use crate::prelude::*;
 use arrow::temporal_conversions::{time64ns_to_time, NANOSECONDS};
 use chrono::Timelike;
-use std::fmt::Write;
 
 pub(crate) fn time_to_time64ns(time: &NaiveTime) -> i64 {
     time.second() as i64 * NANOSECONDS + time.nanosecond() as i64
