+++ polars/polars-core/src/doc/mod.rs
@@ -1,8 +1,2 @@
 //! Other documentation
 pub mod changelog;
-#[cfg(all(
-    feature = "temporal",
-    feature = "dtype-date",
-    feature = "dtype-datetime"
-))]
-pub mod time;
