+++ polars/polars-io/Cargo.toml
@@ -27,7 +27,7 @@ csv-file = ["csv-core", "memmap", "lexical", "arrow/io_csv_write"]
 fmt = ["polars-core/fmt"]
 decompress = ["flate2/miniz_oxide"]
 decompress-fast = ["flate2/zlib-ng-compat"]
-temporal = ["polars-core/dtype-date", "polars-core/dtype-datetime"]
+temporal = ["dtype-datetime", "dtype-date", "dtype-time"]
 # don't use this
 private = []
