+++ polars/polars-time/src/chunkedarray/kernels.rs
@@ -1,5 +1,7 @@
 //! macros that define kernels for extracting
 //! `week`, `weekday`, `year`, `hour` etc. from primitive arrays.
+use super::*;
+use chrono::{Datelike, NaiveDate, NaiveDateTime, Timelike};
 use polars_arrow::export::arrow::array::{ArrayRef, PrimitiveArray};
 use polars_arrow::export::arrow::compute::arity::unary;
 #[cfg(feature = "dtype-time")]
@@ -8,9 +10,7 @@ use polars_arrow::export::arrow::temporal_conversions::{
     date32_to_datetime, timestamp_ms_to_datetime, timestamp_ns_to_datetime,
     timestamp_us_to_datetime,
 };
-use chrono::{Datelike, NaiveDate, NaiveDateTime, Timelike};
 use std::sync::Arc;
-use super::*;
 
 trait PolarsWeekDay {
     fn p_weekday(&self) -> u32;
