+++ polars/polars-time/src/prelude.rs
@@ -1,5 +1,5 @@
+pub use crate::chunkedarray::*;
+pub use crate::series::TemporalMethods;
 pub use crate::windows::{bounds::*, duration::*, groupby::*, window::*};
 pub use crate::*;
 pub use date_range::*;
-pub use crate::series::TemporalMethods;
-pub use crate::chunkedarray::*;
