+++ polars/tests/it/core/mod.rs
@@ -1,1 +1,3 @@
-
+mod list;
+#[cfg(feature = "rows")]
+mod pivot;
